Moved user/channels intersection to rocketController
